purl_sync/requirements.txt
anyio==4.0.0
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
dateparser==1.1.8
Deprecated==1.2.14
Django==4.2.2
django-environ==0.10.0
django-oauth-toolkit==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
http-message-signatures==0.4.4
http-sfv==0.9.8
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
jwcrypto==1.5.0
oauthlib==3.2.2
packageurl-python==0.11.1
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
pycparser==2.21
pytest==7.3.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
saneyaml==0.6.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzlocal==5.0.1
unidiff==0.7.5
urllib3==2.0.3
wrapt==1.15.0
